meta.yaml: give up on tf 2.4 for now
For now, to get the release out the door this week (there are some urgent bugfixes in this release of caiman), I'm giving up on trying to pin tensorflow 2.4+ and figuring out how to get Windows builds to pull that version from defaults rather than conda-forge (which only has very old builds of tensorflow for windows). Nobody seems to know how to do it (although I really appreciate the attempts and help from @isuruf and @beckermr ). I will return to this later.